Fix 9 Tier 1 review issues across new code

+++ src/ui/ToastManager.hpp
@@ -6,15 +6,15 @@ namespace magic {
 
 class ToastManager {
 public:
-    void show(const std::string& message, float duration_sec = 3.0f);
+    void show(const std::string& message, double duration_sec = 3.0);
     void render();
     bool has_pending() const { return !toasts_.empty(); }
 
 private:
     struct Toast {
         std::string message;
-        float start_time;
-        float duration;
+        double start_time;
+        double duration;
     };
     std::deque<Toast> toasts_;
 };
